+++ Locale/enUS.lua
@@ -509,6 +509,7 @@ L["chat_cmd_add_invalid_owner"] = "The player %s was either invalid or not a can
 
 L["ML_ADD_INVALID_ITEM"] = "Invalid itemLink or itemID: %s"
 L["ML_ADD_ITEM_MAX_ATTEMPTS"] = "Couldn't fetch item info for %s - probably not a real item."
+L.ML_REQUEST_VOTES = "%s has requested you to vote for item #%d: %s"
 
 L["error_test_as_non_leader"] = "You cannot initiate a test while in a group without being the group leader."
 L["lootFrame_error_note_required"] = "You must add a note before submitting your response - %s"
@@ -592,3 +593,5 @@ L.opt_useSlashRC_desc = "Disable this if you want to restore '/rc' to ready chec
 L.opt_autoGroupLoot_name = "Auto Group Loot"
 L.opt_autoGroupLoot_desc = "When enabled, RCLootCouncil will automatically click the pass and greed buttons so that all items lands in your inventory."
 
+L.VF_REQUEST_VOTES = "Request Votes"
+L.VF_REQUEST_VOTES_SENT = "Vote request sent"