+++ core.lua
@@ -2377,6 +2377,15 @@ function RCLootCouncil:AddSpecIconToText(specID, text, size)
 	return format("%s %s", specIconCache[id], text)
 end
 
+--- Creates a string with item icon in front of the item link.
+---@param item string|ItemID|ItemLink|ItemString
+---@return string
+function RCLootCouncil:GetItemTextWithIcon(item)
+	local _, itemLink, _, _, _, _, _, _, _, texture = C_Item.GetItemInfo(item)
+	if not texture then return item end -- No icon found, return the item link
+	return format("|T%s:0|t%s", texture, itemLink)
+end
+
 -- cName is name of the module
 function RCLootCouncil:CreateGameTooltip(cName, parent)
 	local itemTooltip = CreateFrame("GameTooltip", cName .. "_ItemTooltip", parent, "GameTooltipTemplate")