+++ changelog.md
@@ -3,6 +3,12 @@
 
 Added a tooltip to any response changed to "Awarded" showing the original response.
 
+### Request Votes
+
+The Master Looter can now request votes from council members on any individual item in session. Doing show will print a message for any council members that have yet to vote for the item.
+
+This can be done either through the right-click menu, or by alt-clicking any vote/unvote button.
+
 # v3.15.5
 
 Updated for patch 11.1.5.