+++ changelog.md
@@ -1,3 +1,4 @@
+# v3.16.0
 
 ### Awarded response